# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Se-test Bruno Luiz Katzjarowski",
        "title": "Se-test Bruno Luiz Katzjarowski",
        "contact": {},
        "license": {},
        "version": "1.0"
    },
    "host": "127.0.0.1:9090",
    "basePath": "/api",
    "paths": {
        "/legacy/": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Get a Negativations by the Customer Document",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "negativations"
                ],
                "summary": "Get a Negativations by the Customer Document",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Customer Document",
                        "name": "company_document",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Negativations"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/types.DefaultMessageReturn"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/types.DefaultMessageReturn"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.DefaultMessageReturn"
                        }
                    }
                }
            }
        },
        "/legacy/start_import": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Get a Negativations by the Customer Document",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "negativations"
                ],
                "summary": "Get a Negativations by the Customer Document",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Customer Document",
                        "name": "company_document",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/types.DefaultMessageReturn"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/types.DefaultMessageReturn"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/types.DefaultMessageReturn"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.DefaultMessageReturn"
                        }
                    }
                }
            }
        },
        "/negativations/company_document/{company_document}": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Get a Negativations by the Company Document",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "negativations"
                ],
                "summary": "Get a Negativations by the Company Document",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Company Document",
                        "name": "company_document",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Negativations"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/types.DefaultMessageReturn"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/types.DefaultMessageReturn"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.DefaultMessageReturn"
                        }
                    }
                }
            }
        },
        "/negativations/customer_document/{customer_document}": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Get all negativations",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "negativations"
                ],
                "summary": "Get all negativations",
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Negativations"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/types.DefaultMessageReturn"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/types.DefaultMessageReturn"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.DefaultMessageReturn"
                        }
                    }
                }
            }
        },
        "/negativations/{id}": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Get a Negativations by the id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "negativations"
                ],
                "summary": "Get a Negativations by the id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Negativation DB id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Negativations"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/types.DefaultMessageReturn"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/types.DefaultMessageReturn"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.DefaultMessageReturn"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Negativations": {
            "type": "object",
            "properties": {
                "companyDocument": {
                    "type": "string"
                },
                "companyName": {
                    "type": "string"
                },
                "contract": {
                    "type": "string"
                },
                "customerDocument": {
                    "type": "string"
                },
                "debtDate": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "inclusionDate": {
                    "type": "string"
                },
                "value": {
                    "type": "number"
                }
            }
        },
        "types.DefaultMessageReturn": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}